Tests are actually unit tests, move accordingly.
